+++ src/database/src/Commands/ModelCommand.php
@@ -99,7 +99,8 @@ class ModelCommand extends Command
             ->setTableMapping($this->getOption('table-mapping', 'commands.gen:model.table_mapping', $pool, []))
             ->setIgnoreTables($this->getOption('ignore-tables', 'commands.gen:model.ignore_tables', $pool, []))
             ->setWithComments($this->getOption('with-comments', 'commands.gen:model.with_comments', $pool, false))
-            ->setVisitors($this->getOption('visitors', 'commands.gen:model.visitors', $pool, []));
+            ->setVisitors($this->getOption('visitors', 'commands.gen:model.visitors', $pool, []))
+            ->setPropertyCase($this->getOption('property-case', 'commands.gen:model.property_case', $pool));
 
         if ($table) {
             $this->createModel($table, $option);
@@ -123,6 +124,7 @@ class ModelCommand extends Command
         $this->addOption('ignore-tables', null, InputOption::VALUE_OPTIONAL | InputOption::VALUE_IS_ARRAY, 'Ignore tables for creating models.');
         $this->addOption('with-comments', null, InputOption::VALUE_NONE, 'Whether generate the property comments for model.');
         $this->addOption('visitors', null, InputOption::VALUE_OPTIONAL | InputOption::VALUE_IS_ARRAY, 'Custom visitors for ast traverser.');
+        $this->addOption('property-case', null, InputOption::VALUE_OPTIONAL, 'Which property case you want use, 0: snake case, 1: camel case.');
     }
 
     protected function getSchemaBuilder(string $poolName): MySqlBuilder